% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_MC_vr.R
\name{CP_MC_vr}
\alias{CP_MC_vr}
\title{Compute the coverage probability of the CIUUPI for a given vector of
\eqn{\gamma} and \eqn{\rho} using Monte Carlo simulation with variance reduction}
\usage{
CP_MC_vr(
  gamma.vec,
  rho,
  N,
  b.knots.vec,
  b1.knots.vec,
  s.knots.vec,
  r.knots.vec,
  t.vec,
  alpha,
  d,
  delta
)
}
\arguments{
\item{gamma.vec}{Vector of \eqn{(\gamma_{p-1}, \gamma_{p})}}

\item{rho}{Vector of \eqn{(\rho_{p-1}, \rho_{p})}}

\item{N}{Number of independent simulation runs}

\item{b.knots.vec}{The vector \eqn{(b(1), b(2), \dots, b(k))}}

\item{b1.knots.vec}{The vector \eqn{(b_1(0), b_1(1), \dots, b_1(k))}}

\item{s.knots.vec}{The vector \eqn{(s(0), s(1), \dots, s(k))}}

\item{r.knots.vec}{The vector \eqn{(r_1(0), r_1(1), \dots, r_1(k))}}

\item{t.vec}{The vector \eqn{(t_0, t_1, \dots, t_k)}}

\item{alpha}{The nominal coverage of the CI is \eqn{1 - \alpha}}

\item{d}{A positive number that specifies the interval \eqn{[-d,d]}}

\item{delta}{A positive number that specifies the interval \eqn{[0,d]}}
}
\value{
The coverage probability computed using Monte Carlo simulation with
variance reduction.
}
\description{
This function computes the coverage probability using Monte Carlo
simulation with variance reduction.
The coverage probability is
\deqn{
{\rm CP}_{\rm MC}(\bm{\gamma}, \bm{\rho})
= \int k(\bm{h}, \bm{\gamma}, \bm{\rho}) \,\bm{\phi}(\bm{h}-\bm{\gamma}) \, d\bm{h}
\\[6pt]
= \text{E}\!\left[ k(\bm{H}, \bm{\gamma}, \bm{\rho}) \right]
}
Here \eqn{\bm{H} \sim N(\bm{\gamma}, \bm{I})} where
\eqn{\bm{H} = (H_{p-1}, H_p)}.
We estimate \eqn{{\rm CP}_{\rm MC}(\bm{\gamma}, \bm{\rho})} by
\deqn{
\frac{1}{N} \sum_{r=1}^{N} k(\bm{h}^r, \bm{\gamma}, \bm{\rho}),
}
where \eqn{N} is the number of independent simulation runs and
\eqn{(h_{p-1}^r, h_p^r)} is the observation of \eqn{\bm{H}} in the
\eqn{r^{th}} simulation run.
}
