% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_T1_mult_int.R
\name{CP_T1_mult_int}
\alias{CP_T1_mult_int}
\title{Compute the coverage probability of the CIUUPI for a given vector of
\eqn{\gamma} and \eqn{\rho} using Theorem 4.5.1}
\usage{
CP_T1_mult_int(
  gamma.vec,
  alpha,
  d,
  rho,
  t.vec,
  s.knots.vec,
  b.knots.vec,
  b1.knots.vec,
  r.knots.vec,
  delta,
  no.nodes.GL
)
}
\arguments{
\item{gamma.vec}{Vector of \eqn{(\gamma_{p-1}, \gamma_{p})}}

\item{alpha}{The nominal coverage of the CI is \eqn{1 - \alpha}}

\item{d}{A positive number that specifies the interval \eqn{[-d,d]}}

\item{rho}{Vector of \eqn{(\rho_{p-1}, \rho_{p})}}

\item{t.vec}{The vector \eqn{(t_0, t_1, \dots, t_k)}}

\item{s.knots.vec}{The vector \eqn{(s(0), s(1), \dots, s(k))}}

\item{b.knots.vec}{The vector \eqn{(b(1), b(2), \dots, b(k))}}

\item{b1.knots.vec}{The vector \eqn{(b_1(0), b_1(1), \dots, b_1(k))}}

\item{r.knots.vec}{The vector \eqn{(r_1(0), r_1(1), \dots, r_1(k))}}

\item{delta}{A positive number that specifies the interval \eqn{[0,d]}}

\item{no.nodes.GL}{The number of Gauss Legendre nodes}
}
\value{
Coverage probability computed using Theorem 4.5.1 and multiple
\code{integrate()} functions.
}
\description{
This function computes the coverage probability using Theorem 4.5.1 and
multiple \code{integrate()} functions.
The coverage probability is
\deqn{
{\rm CP}(\boldsymbol{\gamma}, \boldsymbol{\rho}) = 1 - \alpha +
{\cal I}_1(\bm{\gamma}, \bm{\rho}) + {\cal I}_2(\bm{\gamma}, \bm{\rho})
}
We use multiple \code{integrate()} functions to compute the integrals.
}
