% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_T2_i.R
\name{CP_T2_int}
\alias{CP_T2_int}
\title{Compute the coverage probability of the CIUUPI for a given vector of
\eqn{\gamma} and \eqn{\rho} using Theorem 4.5.2}
\usage{
CP_T2_int(
  rho,
  d,
  alpha,
  gamma.vec,
  t.vec,
  b.knots.vec,
  b1.knots.vec,
  s.knots.vec,
  r.knots.vec,
  delta
)
}
\arguments{
\item{rho}{Vector of \eqn{(\rho_{p-1}, \rho_{p})}}

\item{d}{A positive number that specifies the interval \eqn{[-d,d]}}

\item{alpha}{The nominal coverage of the CI is \eqn{1 - \alpha}}

\item{gamma.vec}{Vector of \eqn{(\gamma_{p-1}, \gamma_{p})}}

\item{t.vec}{The vector \eqn{(t_0, t_1, \dots, t_k)}}

\item{b.knots.vec}{The vector \eqn{(b(1), b(2), \dots, b(k))}}

\item{b1.knots.vec}{The vector \eqn{(b_1(0), b_1(1), \dots, b_1(k))}}

\item{s.knots.vec}{The vector \eqn{(s(0), s(1), \dots, s(k))}}

\item{r.knots.vec}{The vector \eqn{(r_1(0), r_1(1), \dots, r_1(k))}}

\item{delta}{A positive number that specifies the interval \eqn{[0,d]}}
}
\value{
Coverage probability computed using Theorem 4.5.2 and the
\code{integrate()} function.
}
\description{
This function computes the coverage probability using Theorem 4.5.2 and
multiple \code{integrate()} functions.
The coverage probability is
\deqn{
{\rm CP}(\boldsymbol{\gamma}, \boldsymbol{\rho}) = 1 - \alpha - t_1(\gamma_p)
+ t_2(\bm{\gamma}; b, s, b_1, r_1) + t_3(\bm{\gamma}; b, s)
+ t_4(\gamma_p; b, s)
}
We use the \code{integrate()} function to compute the integrals.
}
