% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEL_int.R
\name{SEL_int}
\alias{SEL_int}
\title{Compute the scaled expected length of the CIUUPI for a given vector of
\eqn{\gamma} using the \code{integrate()} function}
\usage{
SEL_int(r.knots.vec, s.knots.vec, t.vec, alpha, d, gamma.vec, delta)
}
\arguments{
\item{r.knots.vec}{The vector \eqn{(r_0, r_1, \dots, r_k)}}

\item{s.knots.vec}{The vector \eqn{(s_0, s_1, \dots, s_k)}}

\item{t.vec}{The vector \eqn{(t_0, t_1, \dots, t_k)}}

\item{alpha}{The nominal coverage of the CI is \eqn{1 - \alpha}}

\item{d}{A positive number that specifies the interval \eqn{[-d,d]}}

\item{gamma.vec}{Vector of \eqn{(\gamma_{p-1}, \gamma_{p})}}

\item{delta}{A positive number that specifies the interval \eqn{[0,d]}}
}
\value{
Scaled expected length computed using the \code{integrate()} function.
}
\description{
Evaluate the scaled expected length of the confidence interval that utilizes
uncertain prior information (CIUUPI) at a given \eqn{\gamma} using the
\code{integrate()} function. In the formula
\deqn{
{\rm SEL}(\bm{\gamma}) =
1 + \frac{1}{ z_{1-\alpha/2}} \int_{-d}^d s(h_p)\,\phi(h_{p} - \gamma_{p})\,dh_p
+ \frac{1}{ z_{1-\alpha/2}} \int_{-\delta}^\delta w_1(|h_p|)\,\phi(h_{p} - \gamma_{p})\,dh_p
  \int_{-d}^{d} r_1(h_{p-1})\,\phi(h_{p-1} - \gamma_{p-1})\,dh_{p-1}
- \big(\Phi(d - \gamma_p) - \Phi(-d - \gamma_p)\big).
}
all integrals are computed using the \code{integrate()} function.
}
