% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEL_MC_vr.R
\name{SEL_MC_vr}
\alias{SEL_MC_vr}
\title{Compute the scaled expected length of the CIUUPI for a given vector of
\eqn{\gamma} using Monte Carlo simulation with variance reduction}
\usage{
SEL_MC_vr(gamma.vec, N, s.knots.vec, r.knots.vec, t.vec, alpha, d, delta)
}
\arguments{
\item{gamma.vec}{Vector of \eqn{(\gamma_{p-1}, \gamma_{p})}}

\item{N}{Number of independent simulation runs}

\item{s.knots.vec}{The vector \eqn{(s_0, s_1, \dots, s_k)}}

\item{r.knots.vec}{The vector \eqn{(r_0, r_1, \dots, r_k)}}

\item{t.vec}{The vector \eqn{(t_0, t_1, \dots, t_k)}}

\item{alpha}{The nominal coverage of the CI is \eqn{1 - \alpha}}

\item{d}{A positive number that specifies the interval \eqn{[-d,d]}}

\item{delta}{A positive number that specifies the interval \eqn{[0,d]}}
}
\value{
The scaled expected length computed using Monte Carlo simulation with
variance reduction.
}
\description{
This function computes the scaled expected length using Monte Carlo
simulation with variance reduction.
The scaled expected length is
\deqn{
{\rm SEL}_{\rm MC}(\bm{\gamma}) = {\rm E}\left(\frac{s(H_p) + w_1(|H_p|) r_1(H_{p-1})}
{ z_{1-\alpha/2}}\right).
}
Here \eqn{\bm{H} \sim N(\bm{\gamma}, \bm{I})} where
\eqn{\bm{H} = (H_{p-1}, H_p)}.
We estimate \eqn{{\rm SEL}_{\rm MC}(\bm{\gamma})} by
\deqn{
\frac{1}{N} \sum_{r=1}^{N} \left(\frac{s(h_p^r) + w_1(|h_p^r|)
r_1(h_{p-1}^r)}{ z_{1-\alpha/2}}\right),
}
where \eqn{N} is the number of independent simulation runs and
\eqn{(h_{p-1}^r, h_p^r)} is the observation of \eqn{\bm{H}} in the
\eqn{r^{th}} simulation run.
}
